add revision

+++ spring-boot-3-jwt-security-main/src/main/java/com/main/repository/TokenRepository.java
@@ -6,15 +6,18 @@ import java.util.Optional;
 import com.main.entity.Token;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.history.RevisionRepository;
 
-public interface TokenRepository extends JpaRepository<Token, Integer> {
+public interface TokenRepository extends JpaRepository<Token, Integer>
+       // , RevisionRepository<Token,Integer,Integer>
+{
 
   @Query(value = """
       select t from Token t inner join User u\s
       on t.user.id = u.id\s
       where u.id = :id and (t.expired = false or t.revoked = false)\s
       """)
-  List<Token> findAllValidTokenByUser(Integer id);
+  List<Token> findAllValidTokenByUser(Long id);
 
   Optional<Token> findByToken(String token);
 }
